requirements.txt
# Ultra PDF Editor - Dependencies
# Core GUI Framework
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0

# PDF Processing (Primary)
PyMuPDF>=1.23.0          # fitz - Fast PDF rendering & manipulation
pypdf>=3.17.0            # PDF merging, splitting, metadata
pikepdf>=8.10.0          # Advanced PDF manipulation

# PDF Generation
reportlab>=4.0.0         # PDF creation from scratch

# Image Processing
Pillow>=10.2.0           # Image manipulation
pdf2image>=1.16.0        # PDF to image conversion

# OCR
pytesseract>=0.3.10      # OCR wrapper for Tesseract

# Document Conversion
python-docx>=1.1.0       # Word document handling
openpyxl>=3.1.0          # Excel file handling
python-pptx>=0.6.23      # PowerPoint handling
beautifulsoup4>=4.12.0   # HTML parsing
lxml>=5.1.0              # XML/HTML processing

# Text/Table Extraction
pdfplumber>=0.10.0       # Text and table extraction
camelot-py[cv]>=0.11.0   # Advanced table extraction

# Security
cryptography>=41.0.0     # Encryption support

# Spell Check
pyenchant>=3.2.0         # Spell checking

# Utilities
natsort>=8.4.0           # Natural sorting
send2trash>=1.8.0        # Safe file deletion
darkdetect>=0.8.0        # System theme detection
